Drop json/md loader bits from common configuration
This was wreaking havoc with `require.context` for some reason. When
Antwar was installed locally, the files were excluded from the result.
This made it impossible to define proper *antwar.config.js*.

Even though this isn't an ideal solution, at least it works now and you
can use Antwar local to a project.